Match MRNs from two EMRs

Takes two data.frame objects with patient information and matches them

match_MRNs(x, y)

Arguments

x, y
Two data frames with columns `mrn`, `first_name`, `last_name`, and `birth_date` (yyyy-mm-dd)

Value

A data.frame with column names with suffixes '_x' and '_y' corresponding to the x and y inputs. The dist column gives the string distance between the two first names.

Details

The function cleans up the last names slightly and then matches exactly on the cleaned up last names and the birth dates. Then fuzzy matching is done on the first name, using the stringdist package.

Examples

data(x_EMR)
data(y_EMR)

match_MRNs(x_EMR, y_EMR)

#>   mrn_x first_name_x last_name_x birth_date_x mrn_y first_name_y last_name_y
#> 1    11        James       Brown   1978-01-01    21        James       Brown
#> 2    12      Charlie      Manson   1952-05-15    22      Charles      Manson
#> 3    14         Gary      Oldman   1960-12-15    24         Gary     Old-man
#> 4    15        Steve       Brown   1980-01-01    25      Stephen       Brown
#> 5    17         Phil      Bowers   1980-01-01    27       Philip   Bowers Jr
#>   birth_date_y dist
#> 1   1978-01-01    0
#> 2   1952-05-15    2
#> 3   1960-12-15    0
#> 4   1980-01-01    3
#> 5   1980-01-01    2